Code for Fused Fiber Photometry
The code allows reproduction of the analysis of data presented in the manuscript "A flexible and versatile system for multi-color fiber photometry and optogenetic manipulation", Cell Method Reports, 2023, by Formozov, Dieter, and Wiegert. 

The corresponding data is deposited at https://gin.g-node.org/SW_lab/FusedFiberPhotometry

Data paths have to be adjusted in the beginning of each script.